Changed freeze scene name
